Cleaned up code, restructured and renamed files
