+++ frontend/src/common/CollapseCard.vue
