+++ frontend/src/common/MapViewer.vue
