+++ frontend/src/common/MenuBar.vue
